Fixing compilation issue. Adding config to package.json Reverting common.ts change.
