+++ package.json
@@ -61,6 +61,7 @@
     "typescript": "^3.6.3",
     "webpack-cli": "^3.3.9",
     "webpack-node-externals": "^1.7.2",
-    "webpack-shell-plugin": "^0.5.0"
+    "webpack-shell-plugin": "^0.5.0",
+    "config": "^3.2.2"
   }
 }
